@@ -1,4 +1,4 @@
-class AddAuthenticationTokenToUsers < ActiveRecord::Migration
+class AddAuthenticationTokenToUsers < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :authentication_token, :string
   end
